+++ src/index.ts
@@ -1,9 +1,9 @@
 export { analyzePGN, printHeatmap } from '#core/analyze';
 export { GameTrackerBase, MoveTracker } from '#tracker/base-tracker';
 export { default as BaseTracker } from '#tracker/base-tracker';
-export { default as PieceTracker, isTrackedPiece } from '#tracker/piece-tracker-base';
-export { default as GameTracker } from '#tracker/game-tracker-base';
-export { default as TileTracker } from '#tracker/tile/tile-tracker-base';
+export { default as PieceTracker, isTrackedPiece } from '#tracker/piece-tracker';
+export { default as GameTracker } from '#tracker/game-tracker';
+export { default as TileTracker } from '#tracker/tile/tile-tracker';
 export type { Action } from '#types/actions';
 export type {
     AnalyzeOptions,